refactor: move `GetEffectiveObjectsForLayerMapping` to `RevitMappingResolver`
